SuperStore Sales agent Dashboard
Superstore is a famous retailer in Canada. They have expanded into the USA and their
business model involves placing Sales Agents in every state in the USA. These sales agents are
responsible for bringing in sales for the state that they are assigned to.
The Sales Manager wants a dashboard to track how Sales Agents are doing. He wants it to be
comparable where he selects a timeframe this year (Example January - March 2020) and he
wants sales data for the same period last year. This is so he can track growth.
Additionally, he wants a breakdown of sales per category and subcategory as well as the top
performing products.